Refactor Redis elements out into helpers_redis module
